requirements.txt
absl-py==0.13.0
alembic==1.6.5
aniso8601==7.0.0
asgiref==3.3.4
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
bcrypt==3.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
deap==1.3.1
dill==0.3.4
Django==3.2.4
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
huggingface-hub==0.0.12
idna==2.10
iniconfig==1.1.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jupyter-core==4.7.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
names==0.3.0
numpy==1.20.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
pandas==1.3.0
parso==0.8.2
patsy==0.5.1
pep517==0.10.0
pickleshare==0.7.5
Pillow==8.2.0
pip-tools==6.2.0
pluggy==0.13.1
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.17.3
psycopg2==2.9.1
psycopg2-binary==2.8.6
ptvsd==4.3.2
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
python-dotenv-run==0.1.4
python-editor==1.0.4
pytz==2021.1
pywin32==301
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Rx==1.6.1
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
selenium==3.141.0
seqeval==1.2.2
six==1.16.0
SQLAlchemy==1.4.17
sqlparse==0.4.1
statsmodels==0.12.2
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.1.0
tokenizers==0.10.3
toml==0.10.2
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
transformers==4.8.1
typing-extensions==3.10.0.0
urllib3==1.26.5
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
